fix touchboard stuck key bug
Sometimes touch events come so rapidly that the touch handlers try to
"release" a note before it even has a chance to start properly. The
PlayingNote constructor calls its start method in a setTimeout, and that
usually sets up some stuff to be called on release. But the touch
handlers were calling release before that setTimeout came back, and thus
before the PlayingNote was set up to handle a release. Fixed by making
sure the touch handlers call release inside a setTimeout as well, so
that it gets queued behind the one from the constructor.

While debugging this I also fixed some other minor issues with the touch
handlers and various kinds of missing values.